NZSL-51: Refactor Freelex code into a module, add bridge class to resolve correct class (#1476)
* Add dependency on sqlite3 to load NZSL Signbank export dictionary

* Add setting to specify the dictionary adapter

* Move Freelex code into a module so that we can add Signbank equivalents alongside

* Add SignBank model to bridge call sites that work outside the Freelex module to the correct sign model based on app config

* Resolve rubocop warnings blocking passing build

* Apply comment and variable name update suggestions from code review


* Fix syntax error in Freelex::SignMenu
