+++ spec/features/home_spec.rb
@@ -10,7 +10,7 @@ describe 'Test with visual testing', type: :feature, js: true do
 
   before do
     SeedDataService.load_all
-    allow(Sign).to receive(:random).and_return Sign.first(id: sign_id)
+    allow(SignModel.resolve).to receive(:random).and_return SignModel.resolve.first(id: sign_id)
   end
   describe 'home' do
     it 'root_path' do
