+++ spec/spec_helper.rb
@@ -88,4 +88,6 @@ RSpec.configure do |config|
   # # test failures related to randomization by passing the same `--seed` value
   # # as the one that triggered the failure.
   # Kernel.srand config.seed
+
+  config.example_status_persistence_file_path = './tmp/rspec.txt'
 end
